prevent backup for dry-run

+++ src/backup.cpp
@@ -24,6 +24,9 @@ struct BackupPrivate
 Backup::Backup()
     : d(new BackupPrivate)
 {
+    if (ArgumentsAndSettings::dryRun())
+        return;
+
     if (ArgumentsAndSettings::backupDir().isEmpty()) {
         QTemporaryDir dir;
         if (dir.isValid()) {
@@ -41,13 +44,19 @@ Backup::Backup()
 
 Backup::~Backup()
 {
-    if (d->tempBackup)
-        destroy();
+    if (!ArgumentsAndSettings::dryRun()) {
+        if (d->tempBackup)
+            destroy();
+    }
+
     delete d;
 }
 
 bool Backup::backupOneFile(const QString &pathRelativeToQtDir_)
 {
+    if (ArgumentsAndSettings::dryRun())
+        return false;
+
     QString pathRelativeToQtDir = QDir::cleanPath(pathRelativeToQtDir_);
     // detect possibly erroneous backup operation?
     if (pathRelativeToQtDir.contains(QStringLiteral(".."))) {
@@ -68,6 +77,9 @@ bool Backup::backupOneFile(const QString &pathRelativeToQtDir_)
 
 bool Backup::restoreAll()
 {
+    if (ArgumentsAndSettings::dryRun())
+        return false;
+
     foreach (const QString &file, d->filesMadeBackup) {
         QBPLOGV(QString(QStringLiteral("restoring backup file %1")).arg(file));
         QFile::copy(d->backupDir.absoluteFilePath(file), d->qtDir.absoluteFilePath(file));
@@ -79,6 +91,9 @@ bool Backup::restoreAll()
 
 void Backup::destroy()
 {
+    if (ArgumentsAndSettings::dryRun())
+        return;
+
     QBPLOGV(QString(QStringLiteral("deleting backup dir %1")).arg(d->backupDir.absolutePath()));
     d->backupDir.removeRecursively();
     d->filesMadeBackup.clear();
